Use fixtures for examples spec
Up until now, the tests ran against the real `govuk-content-schemas`
repository. This was ok, but we'd now like to evolve the folder structure and
we would like to be able to test this in isolation from the real files. This
will allow us to update this gem to be compatible with both folder structures,
and so allow us to iterate towards the new structure safely.

It's also generally good practice to have tests isolated from other factors
which may change.