Clean up specs and Validator
The `Validator` did a lot of work because it had to look inside two
types of directories
(https://github.com/alphagov/govuk-content-schema-test-helpers/pull/14).
 This was done to facilitate migration between the content-schemas
structures.

That migration has been finished, and we apps won't be relying on the
old structure (if they ever did). Removing the feature will allow us to
clean up a bunch of code.